wip : dof 기본적인것 구현. 사용성 개선 필요

+++ EngineSIU/EngineSIU/Engine/Source/Runtime/Renderer/BillboardRenderPass.h
@@ -6,7 +6,7 @@
 
 #include "Define.h"
 
-enum class EResourceType : uint16;
+enum class EResourceType : uint8;
 class UBillboardComponent;
 class FDXDBufferManager;
 class FGraphicsDevice;
